embeddings/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
blis==1.2.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
dill==0.3.8
distro==1.9.0
editdistance==0.8.1
einops==0.8.0
exceptiongroup==1.2.2
fasttext==0.9.3
filelock==3.17.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.60.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.42.4
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
